+++ src/main/java/org/example/springboot25/controller/CatGraphQLController.java
@@ -2,9 +2,9 @@ package org.example.springboot25.controller;
 
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
-import org.example.springboot25.DTO.CatInputDTO;
-import org.example.springboot25.DTO.CatOutputDTO;
-import org.example.springboot25.DTO.CatUpdateDTO;
+import org.example.springboot25.dto.CatInputDTO;
+import org.example.springboot25.dto.CatOutputDTO;
+import org.example.springboot25.dto.CatUpdateDTO;
 import org.example.springboot25.entities.Cat;
 import org.example.springboot25.exceptions.CatNotFoundException;
 import org.example.springboot25.mapper.CatMapper;
