+++ src/main/java/org/example/springboot25/dto/CatInputDTO.java
@@ -1,13 +1,25 @@
-package org.example.springboot25.DTO;
+package org.example.springboot25.dto;
 
-public class CatUpdateDTO {
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
+public class CatInputDTO {
+
+    @NotBlank
     private String catName;
+
     private String catProfilePicture;
+
     private String catBreed;
+
     private String catGender;
+
     private Integer catAge;
+
     private String catPersonality;
+
+    @NotNull
+    private Long userId;
 
     public String getCatName() {
         return catName;
@@ -55,5 +67,13 @@ public class CatUpdateDTO {
 
     public void setCatPersonality(String catPersonality) {
         this.catPersonality = catPersonality;
+    }
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
     }
 }
