+++ src/main/java/org/example/springboot25/dto/CatUpdateDTO.java
@@ -1,25 +1,13 @@
-package org.example.springboot25.DTO;
+package org.example.springboot25.dto;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
+public class CatUpdateDTO {
 
-public class CatInputDTO {
-
-    @NotBlank
     private String catName;
-
     private String catProfilePicture;
-
     private String catBreed;
-
     private String catGender;
-
     private Integer catAge;
-
     private String catPersonality;
-
-    @NotNull
-    private Long userId;
 
     public String getCatName() {
         return catName;
@@ -67,13 +55,5 @@ public class CatInputDTO {
 
     public void setCatPersonality(String catPersonality) {
         this.catPersonality = catPersonality;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
     }
 }
